% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{biomarkers}
\alias{biomarkers}
\alias{biomarkers,MOBResult-method}
\title{Accessor for biomarker table in a MOBResult}
\usage{
biomarkers(x, ...)

\S4method{biomarkers}{MOBResult}(x, ...)
}
\arguments{
\item{x}{A \code{MOBResult} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{DataFrame} with columns \code{feature},
  \code{omics_layer}, \code{loading_score}, \code{rank}, and
  \code{component}.
}
\description{
Returns the ranked multi-omics biomarker \code{DataFrame}
  from a \code{MOBResult} object.
}
\examples{
library(S4Vectors)
bm <- DataFrame(feature = c("G1","T1"), omics_layer = c("host","microbiome"),
                loading_score = c(0.8,0.6), rank = c(1L,2L), component = c(1L,1L))
obj <- MOBResult(matrix(rnorm(20), 10, 2), list(), bm, list())
biomarkers(obj)

}
