% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{featureLoadings}
\alias{featureLoadings}
\alias{featureLoadings,MOBResult-method}
\title{Accessor for feature loadings in a MOBResult}
\usage{
featureLoadings(x, ...)

\S4method{featureLoadings}{MOBResult}(x, ...)
}
\arguments{
\item{x}{A \code{MOBResult} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A named \code{list} with elements \code{host} and
  \code{microbiome}, each a matrix of genes/taxa x components.
}
\description{
Returns the named list of per-layer feature loading
  matrices from a \code{MOBResult} object.
}
\examples{
library(S4Vectors)
fl <- list(host = matrix(rnorm(10), 5, 2), microbiome = matrix(rnorm(6), 3, 2))
obj <- MOBResult(matrix(rnorm(20), 10, 2), fl, DataFrame(), list())
featureLoadings(obj)

}
