% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{integrationScores}
\alias{integrationScores}
\alias{integrationScores,MOBResult-method}
\title{Accessor for integrated sample scores in a MOBResult}
\usage{
integrationScores(x, ...)

\S4method{integrationScores}{MOBResult}(x, ...)
}
\arguments{
\item{x}{A \code{MOBResult} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{matrix} with rows = samples and columns = latent
  components.
}
\description{
Returns the matrix of integrated sample scores (samples x
  latent components) from a \code{MOBResult} object.
}
\examples{
library(S4Vectors)
scores <- matrix(rnorm(20), nrow = 10, ncol = 2,
                 dimnames = list(paste0("S",1:10), c("Comp1","Comp2")))
obj <- MOBResult(scores, list(), DataFrame(), list())
integrationScores(obj)

}
