% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{MOBResult}
\alias{MOBResult}
\title{Constructor for MOBResult}
\usage{
MOBResult(
  integratedScores,
  featureLoadings,
  biomarkerTable,
  classifierResults,
  params = list()
)
}
\arguments{
\item{integratedScores}{A \code{matrix} of integrated sample scores
(samples x components).}

\item{featureLoadings}{A named \code{list} with \code{host} and
\code{microbiome} loading matrices.}

\item{biomarkerTable}{A \code{DataFrame} of ranked multi-omics
biomarkers.}

\item{classifierResults}{A named \code{list} of classifier performance
metrics.}

\item{params}{A \code{list} of analysis parameters.}
}
\value{
A \code{MOBResult} object.
}
\description{
Create a new \code{MOBResult} object.
}
\examples{
library(S4Vectors)
scores <- matrix(rnorm(20), nrow = 10, ncol = 2,
                 dimnames = list(paste0("S", 1:10), c("Comp1","Comp2")))
loadings <- list(
    host       = matrix(rnorm(10), nrow = 5, ncol = 2,
                        dimnames = list(paste0("G", 1:5), c("Comp1","Comp2"))),
    microbiome = matrix(rnorm(6), nrow = 3, ncol = 2,
                        dimnames = list(paste0("T", 1:3), c("Comp1","Comp2")))
)
bm <- DataFrame(
    feature      = c("G1", "T1"),
    omics_layer  = c("host", "microbiome"),
    loading_score = c(0.8, 0.6),
    rank         = c(1L, 2L),
    component    = c(1L, 1L)
)
obj <- MOBResult(
    integratedScores  = scores,
    featureLoadings   = loadings,
    biomarkerTable    = bm,
    classifierResults = list(),
    params = list(integration_method = "DIABLO")
)
obj

}
